Added minipad.py updates to macro1.py
Uses midi to interact with vb buttons and update their status.

Needs reworking but for now it's doing everything I want.

Created updates class for updates to db and macro buttons.

Added some logic to functions oo and oi. They will be used for updating the macro buttons even though the track SOLO is done in the DAW.